+++ imod/tests/test_mf6/test_circle.py
@@ -80,7 +80,7 @@ def test_gwfmodel_render(circle_model, tmp_path):
     assert actual == expected
     validation_context = ValidationContext(True)
     write_context2 = WriteContext(tmp_path)
-    gwfmodel.write("GWF_1", globaltimes, write_context2, validation_context)
+    gwfmodel._write("GWF_1", globaltimes, write_context2, validation_context)
     assert (tmp_path / "GWF_1" / "GWF_1.nam").is_file()
     assert (tmp_path / "GWF_1").is_dir()
 
@@ -135,6 +135,6 @@ def test_gwfmodel_render_evt(circle_model_evt, tmp_path):
     assert actual == expected
     validation_context = ValidationContext(True)
     write_context2 = WriteContext(tmp_path)
-    gwfmodel.write("GWF_1", globaltimes, write_context2, validation_context)
+    gwfmodel._write("GWF_1", globaltimes, write_context2, validation_context)
     assert (tmp_path / "GWF_1" / "GWF_1.nam").is_file()
     assert (tmp_path / "GWF_1").is_dir()