+++ docs/api/changelog.rst
@@ -71,6 +71,9 @@ Changed
 - :func:`imod.prepare.fill` now takes a ``dims`` argument instead of ``by``,
   and will fill over N dimensions. Secondly, the function no longer takes
   an ``invalid`` argument, but instead always treats NaNs as missing.
+- Reverted the need for providing WriteContext objects to MODFLOW6 Model and
+  Package objects' ``write`` method. These now use similar arguments to the
+  :meth:`imod.mf6.Modflow6Simulation.write` method.
 - :class:`imod.msw.CouplingMapping`, :class:`imod.msw.Sprinkling`,
   `imod.msw.Sprinkling.MetaSwapModel`, now take the
   :class:`imod.mf6.mf6_wel_adapter.Mf6Wel` and the